New name, get config from multiple places, remove example.
I can't decide on a good name, some that I thought of are taken. I'm
going with Jack Of All Trades (joat) for now as it is not taken and is
short. I read on wikipedia that it can be derrogative but the
derrogative part is kind of true. Joat doesn't know about anything, any
extension of joat will know about the REST API but the end game is users
further specifying scripts on top of joat extensions.

For now it takes config from two places the home directory and local.
In local it will try first .<extension name>.joat/<extension name>.yml
and if that doesn't exist it tries <extension name>.yml. On home it
tries to use ~/.<extension name>.joat/<extension name>.yml.

I haven't decided yet what overrides what in the config files, now home
config ovewrites the local, this is interesting because final users will
not be able to overwrite subcommands. Which is an arguable advantage
since one might want to do that, but in that case the user would lose
the original subcommand and potentially have to reimplement it with a
different name.  Joat could be smarter about merging yaml configs too,
but that can wait.